Transfer selectors from Main component to selectors.js
