+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import {createStore} from 'redux';
 import {Provider} from 'react-redux';
 import App from './components/app/app';
-import {reducer} from './reducer';
+import {reducer} from './reducer/reducer';
 
 const store = createStore(
     reducer,
